Structure the code into modules and packages
* Do not use objects from top layer in bottom
