+++ applitools/core/__init__.py
@@ -1,0 +1,22 @@
+from .triggers import *
+from .test_results import *
+from .target import *
+from .match_window_task import *
+from .logger import *
+from .errors import *
+# from .capture import *
+from .scaling import *
+from .eyes_base import *
+from .geometry import *
+
+__all__ = (triggers.__all__ +
+           test_results.__all__ +
+           match_window_task.__all__ +
+           logger.__all__ +
+           errors.__all__ +
+           scaling.__all__ +
+           # capture.__all__ +
+           eyes_base.__all__ +
+           geometry.__all__ +
+           ('logger',)
+           )
