+++ applitools/core/agent_connector.py
@@ -6,13 +6,13 @@ import typing as tp
 import requests
 from requests.packages import urllib3
 
-from . import logger
-from .test_results import TestResults
-from .utils import general_utils
+from applitools.core import logger
+from applitools.core.test_results import TestResults
+from applitools.utils import general_utils
 
 if tp.TYPE_CHECKING:
     from requests.models import Response
-    from .utils._custom_types import RunningSession, SessionStartInfo, Num
+    from ..utils.custom_types import RunningSession, SessionStartInfo, Num
 
 # Prints out all data sent/received through 'requests'
 # import httplib
