+++ applitools/selenium/__init__.py
@@ -1,0 +1,8 @@
+from .capture import EyesScreenshot
+from .eyes import Eyes
+from .webdriver import EyesWebDriver, EyesFrame
+from .webelement import EyesWebElement
+
+__all__ = (
+        ('Eyes', 'EyesWebElement', 'EyesWebDriver', 'EyesFrame', 'EyesScreenshot')
+)
