+++ applitools/utils/__init__.py
@@ -1,9 +1,7 @@
-import glob
+from .compat import ABC, range, iteritems
+from . import image_utils, argument_guard
+from .general_utils import cached_property
 
-modules = glob.glob('*.py')
-try:
-    modules.remove('__init__.py')
-except ValueError:
-    pass
-
-__all__ = modules
+__all__ = (compat.__all__ +
+           ('image_utils', 'argument_guard')
+           )
