+++ applitools/utils/custom_types.py
@@ -7,7 +7,8 @@ if tp.TYPE_CHECKING:
     from selenium.webdriver.remote.webdriver import WebDriver
     from selenium.webdriver.remote.webelement import WebElement
 
-    from applitools._webdriver import EyesWebDriver, EyesWebElement
+    from applitools.selenium.webdriver import EyesWebDriver
+    from applitools.selenium.webelement import EyesWebElement
 
     RunningSession = tp.Dict[tp.Text, tp.Any]
     ViewPort = tp.Dict[tp.Text, int]
