+++ tests/test_client_sites.py
@@ -1,7 +1,7 @@
 import pytest
 from selenium.webdriver.common.by import By
 
-from applitools.target import Target
+from applitools.core.target import Target
 
 
 @pytest.mark.skip("Depending on Fluent API. Not implemented yet")
